MySQL: Reduce size of counters for global/hood stats
This is just a change of the database!

+++ ffmap/db/stats.py
@@ -11,10 +11,10 @@ mysql = FreifunkMySQL()
 mysql.execute("""
 	CREATE TABLE `stats_global` (
 		`time` int(11) NOT NULL,
-		`clients` int(11) NOT NULL,
-		`online` int(11) NOT NULL,
-		`offline` int(11) NOT NULL,
-		`unknown` int(11) NOT NULL
+		`clients` mediumint(9) NOT NULL,
+		`online` smallint(6) NOT NULL,
+		`offline` smallint(6) NOT NULL,
+		`unknown` smallint(6) NOT NULL
 	) ENGINE=InnoDB DEFAULT CHARSET=utf8 COLLATE=utf8_unicode_ci
 """)
 
@@ -27,10 +27,10 @@ mysql.execute("""
 	CREATE TABLE stats_hood (
 		`hood` varchar(50) CHARACTER SET utf8 COLLATE utf8_bin NOT NULL,
 		`time` int(11) NOT NULL,
-		`clients` int(11) NOT NULL,
-		`online` int(11) NOT NULL,
-		`offline` int(11) NOT NULL,
-		`unknown` int(11) NOT NULL
+		`clients` mediumint(9) NOT NULL,
+		`online` smallint(6) NOT NULL,
+		`offline` smallint(6) NOT NULL,
+		`unknown` smallint(6) NOT NULL
 	) ENGINE=InnoDB DEFAULT CHARSET=utf8 COLLATE=utf8_unicode_ci
 """)
